+++ lib/src/models/theme/chart_theme.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/material.dart';
 import 'package:flutter_chart/src/models/axis/axis_builder.dart';
 import 'package:flutter_chart/src/models/raster/raster_style.dart';
 
@@ -9,8 +10,24 @@ class ChartTheme {
   /// Raster style
   final RasterStyle rasterStyle;
 
+  /// The height of the xAxis where the labels are drawn
+  final double xAxisHeight;
+
+  /// The width of the yAxis where the labels are drawn
+  final double yAxisWidth;
+
+  /// The padding right of the yAxis where the labels are drawn
+  final double yAxisPaddingRight;
+
+  /// Axis text style used for calculating the size of the labels
+  final TextStyle? axisTextStyle;
+
   const ChartTheme({
+    this.xAxisHeight = 30,
+    this.yAxisWidth = 30,
+    this.yAxisPaddingRight = 0,
     this.axisBuilder,
+    this.axisTextStyle,
     this.rasterStyle = const RasterStyle(),
   });
 }