+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+- Added option to modify xAxis height and yAxis width
+- Added option to specify Axis label font size
+- Exported calculateTextSize function
+- Improve point detection logic
+
 ## 0.0.1
 
 - Intial release